+++ Common/src/main/java/de/markusbordihn/fireextinguisher/blockitem/FireAlarmSmokeDetectorBlockItem.java
@@ -1,0 +1,77 @@
+/*
+ * Copyright 2023 Markus Bordihn
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package de.markusbordihn.fireextinguisher.blockitem;
+
+import de.markusbordihn.fireextinguisher.Constants;
+import java.util.List;
+import javax.annotation.Nullable;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Block;
+
+public class FireAlarmSmokeDetectorBlockItem extends BlockItem {
+
+  public static final String NAME = "fire_alarm_smoke_detector";
+
+  public FireAlarmSmokeDetectorBlockItem(Block block) {
+    this(block, new Properties());
+  }
+
+  public FireAlarmSmokeDetectorBlockItem(Block block, Properties properties) {
+    super(block, properties);
+  }
+
+  @Override
+  public void appendHoverText(
+      ItemStack itemStack,
+      @Nullable Level level,
+      List<Component> tooltipList,
+      TooltipFlag tooltipFlag) {
+    tooltipList.add(
+        new TranslatableComponent(Constants.TOOLTIP_PREFIX + NAME).withStyle(ChatFormatting.GRAY));
+    tooltipList.add(
+        new TranslatableComponent(Constants.TOOLTIP_PREFIX + "fire_alarm_smoke_detector.status")
+            .withStyle(ChatFormatting.GRAY));
+    tooltipList.add(
+        new TextComponent("⏹ ")
+            .withStyle(ChatFormatting.YELLOW)
+            .append(
+                new TranslatableComponent(
+                    Constants.TOOLTIP_PREFIX + "fire_alarm_smoke_detector.status.orange")));
+    tooltipList.add(
+        new TextComponent("⏹ ")
+            .withStyle(ChatFormatting.GREEN)
+            .append(
+                new TranslatableComponent(
+                    Constants.TOOLTIP_PREFIX + "fire_alarm_smoke_detector.status.green")));
+    tooltipList.add(
+        new TextComponent("⏹ ")
+            .withStyle(ChatFormatting.RED)
+            .append(
+                new TranslatableComponent(
+                    Constants.TOOLTIP_PREFIX + "fire_alarm_smoke_detector.status.red")));
+  }
+}