+++ Forge/src/main/java/de/markusbordihn/fireextinguisher/block/ModBlocks.java
@@ -31,7 +31,29 @@ public class ModBlocks {
       DeferredRegister.create(ForgeRegistries.BLOCKS, Constants.MOD_ID);
   // Fire Extinguisher
   public static final RegistryObject<Block> FIRE_EXTINGUISHER =
-      BLOCKS.register(FireExtinguisherBlock.NAME, FireExtinguisherBlock::new);
+      BLOCKS.register(FireExtinguisherBlock.NAME, () -> FireExtinguisherBlocks.FIRE_EXTINGUISHER);
+  public static final RegistryObject<Block> FIRE_EXTINGUISHER_COPPER =
+      BLOCKS.register(
+          FireExtinguisherBlock.NAME + "_copper",
+          () -> FireExtinguisherBlocks.FIRE_EXTINGUISHER_COPPER);
+
+  // Fire Alarm
+  public static final RegistryObject<Block> FIRE_SPRINKLER =
+      BLOCKS.register(FireSprinklerBlock.NAME, () -> FireAlarmBlocks.FIRE_SPRINKLER);
+  public static final RegistryObject<Block> FIRE_ALARM_SWITCH =
+      BLOCKS.register(FireAlarmSwitchBlock.NAME, () -> FireAlarmBlocks.FIRE_ALARM_SWITCH);
+  public static final RegistryObject<Block> FIRE_ALARM_BELL =
+      BLOCKS.register(FireAlarmBellBlock.NAME, () -> FireAlarmBlocks.FIRE_ALARM_BELL);
+  public static final RegistryObject<Block> FIRE_ALARM_SIREN =
+      BLOCKS.register(FireAlarmSirenBlock.NAME, () -> FireAlarmBlocks.FIRE_ALARM_SIREN);
+  public static final RegistryObject<Block> FIRE_ALARM_SMOKE_DETECTOR =
+      BLOCKS.register(
+          FireAlarmSmokeDetectorBlock.NAME, () -> FireAlarmBlocks.FIRE_ALARM_SMOKE_DETECTOR);
+  public static final RegistryObject<Block> FIRE_ALARM_SMOKE_DETECTOR_SILENT =
+      BLOCKS.register(
+          FireAlarmSmokeDetectorSilentBlock.NAME,
+          () -> FireAlarmBlocks.FIRE_ALARM_SMOKE_DETECTOR_SILENT);
+
   // Fire Extinguisher Signs
   public static final RegistryObject<Block> FIRE_EXTINGUISHER_SIGN =
       BLOCKS.register(