+++ CHANGELOG.md
@@ -5,6 +5,17 @@
 This change log includes the summarized changes.
 For the full changelog, please go to the [GitHub History][history] instead.
 
+### 6.1.0
+
+- Added Fire Sprinkler Block
+- Added Fire Alarm Bell Block
+- Added Fire Alarm Siren Block
+- Added Fire Alarm Switch Block
+- Added Smoke Detector Block
+- Added Smoke Detector (silent) Block
+- Added Copper Fire Extinguisher Item
+- Improved textures and models
+
 ### 6.0.0
 
 - Refactored code and improved performance